Fix buffer size when using bigger input values
Before, the max/min was relative to the original input. Now, after
normalization, the max/min is calculated again to allocate only what
we truly need.

The program now will work with bigger values (any amount of digits), as
long as the numbers are similar in magnitude. What still doesn't work
if we have big values, that are very far apart.
      Now Working: 1000002, 1000006, 1000009
Still not working: 1000002, 5000006, 8000009